fix gosec G115 linter

+++ pkg/webhook/server/server.go
@@ -57,11 +57,11 @@ var (
 type Server struct {
 	// ListenAddr is the address the HTTP server should listen on
 	// This must be specified.
-	ListenAddr int32
+	ListenAddr int
 
 	// HealthzAddr is the address the healthz HTTP server should listen on
 	// If not specified, the healthz endpoint will not be exposed.
-	HealthzAddr *int32
+	HealthzAddr *int
 
 	// PprofAddress is the address the pprof endpoint should be served on if enabled.
 	PprofAddress string
@@ -134,7 +134,7 @@ func (s *Server) Run(ctx context.Context) error {
 			return err
 		}
 
-		s.ListenAddr = int32(webhookPort)
+		s.ListenAddr = webhookPort
 	}
 
 	mgr, err := ctrl.NewManager(
@@ -155,7 +155,7 @@ func (s *Server) Run(ctx context.Context) error {
 				},
 			},
 			WebhookServer: webhook.NewServer(webhook.Options{
-				Port: int(s.ListenAddr),
+				Port: s.ListenAddr,
 				TLSOpts: []func(*tls.Config){
 					func(cfg *tls.Config) {
 						cfg.CipherSuites = cipherSuites
@@ -289,7 +289,7 @@ func (s *Server) Port() (int, error) {
 		return 0, ErrNotListening
 	}
 
-	return int(s.ListenAddr), nil
+	return s.ListenAddr, nil
 }
 
 func (s *Server) handleHealthz(w http.ResponseWriter, req *http.Request) {
